refactor(shell): move graph-product to graph-shell, group its multiplayer files
Completes the directory half of the rename for the graph adapter, and
gathers the three files that know about rooms into one place:

  graph-shell/
    context.ts  index.ts  shortcuts.ts  types.ts  useGraphShell.ts
    multiplayer/
      bindGraphToDoc.ts  bindGraphToDoc.test.ts  trackDraggedNodes.ts

bindGraphToDoc and trackDraggedNodes are the only files here that reach
for @multiplayer/protocol or yjs, so the subdirectory draws the line
where the dependency already did.

The @magic/shared/graph-product export subpath becomes
@magic/shared/graph-shell, and the two Yjs transaction origin symbols are
relabelled to match their new home.
