+++ packages/magic-products/src/dev/ActionBar.vue
@@ -5,7 +5,7 @@
   import Node from '@magic/shared/Node';
   import VStack from '@magic/shared/VStack';
   import Well from '@magic/shared/Well';
-  import { useProvidedGraph } from '@magic/shared/graph-product';
+  import { useProvidedGraph } from '@magic/shared/graph-shell';
   import { useFocusedNode } from '@magic/shared/utilities';
 
   const graph = useProvidedGraph();
