+++ packages/magic-products/src/markov-chains/MainView.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
   import Shell from '@magic/shared/Shell';
-  import { useGraphShell } from '@magic/shared/graph-product';
+  import { useGraphShell } from '@magic/shared/graph-shell';
 
   const graph = useGraphShell({
     productId: 'markov-chains',
