+++ packages/magic-products/src/min-spanning-trees/simulations/components/EdgeListPanel.vue
@@ -4,7 +4,7 @@
   import Node from '@magic/shared/Node';
   import VStack from '@magic/shared/VStack';
   import Well from '@magic/shared/Well';
-  import { useProvidedGraph } from '@magic/shared/graph-product';
+  import { useProvidedGraph } from '@magic/shared/graph-shell';
   import type { GEdge } from '@magic/shared/graph/types';
 
   import { computed } from 'vue';
