+++ packages/magic-products/src/path-finding/TransitionMatrix.vue
@@ -5,7 +5,7 @@
   import VStack from '@magic/shared/VStack';
   import Well from '@magic/shared/Well';
   import { GNode } from '@magic/shared/graph';
-  import { useProvidedGraph } from '@magic/shared/graph-product';
+  import { useProvidedGraph } from '@magic/shared/graph-shell';
 
   import { computed } from 'vue';
