+++ packages/magic-products/src/traversals/lensChips.ts
@@ -1,4 +1,4 @@
-import { GraphLensChipOption } from '@magic/shared/graph-product';
+import { GraphLensChipOption } from '@magic/shared/graph-shell';
 
 export const lensChips: GraphLensChipOption = (graph) => {
   return;
