+++ packages/magic-shared/src/graph-shell/context.ts
