+++ packages/magic-shared/src/graph-shell/index.ts
