+++ packages/magic-shared/src/graph-shell/shortcuts.ts
