+++ packages/magic-shared/src/graph-shell/types.ts
